rename files according to assignment instructions
