+++ warzone/Strategies.cpp
@@ -1,4 +1,4 @@
-#include "Strategy.h";
+#include "Strategies.h";
 
 /// <summary>
 /// Default constructor for PlayerStrategy
